requirements.txt
fastapi==0.95.2
uvicorn[standard]
databases==0.7.0
greenlet==2.0.2
SQLAlchemy==1.4.49
typing_extensions==4.7.1
asyncpg==0.28.0
psycopg2==2.9.6
python-dotenv==1.0.0